<feat> update doc swagger and create Schema for the AuthController
