+++ backend/app/Docs/Schemas/AuthErrorSchema.php
@@ -1,0 +1,16 @@
+<?php
+namespace App\Docs\Schemas;
+
+/**
+ * @OA\Schema(
+ *     schema="AuthError",
+ *     type="object",
+ *     title="Error de autenticación",
+ *     @OA\Property(
+ *         property="error",
+ *         type="string",
+ *         example="Credenciales inválidas"
+ *     )
+ * )
+ */
+class AuthErrorSchema {}
