+++ backend/app/Docs/Schemas/UnauthorizedSchema.php
@@ -1,0 +1,16 @@
+<?php
+namespace App\Docs\Schemas;
+
+/**
+ * @OA\Schema(
+ *     schema="Unauthorized",
+ *     type="object",
+ *     title="No autorizado",
+ *     @OA\Property(
+ *         property="message",
+ *         type="string",
+ *         example="No autorizado"
+ *     )
+ * )
+ */
+class UnauthorizedSchema {}
